[CHANGED] slider的数值可以用小数表示，不再用int后缩放

+++ easyparticle2dnew/src/libparticle2d/Symbol.cpp
@@ -79,19 +79,19 @@ void Symbol::loadResources()
 	fin.close();
 
  	m_ps = new ParticleSystem(1000);
-	m_ps->SetValue(PS_EMISSION_TIME, UICallback::Data(value[ITEM_EMISSION_TIME].asInt()));
+	m_ps->SetValue(PS_EMISSION_TIME, UICallback::Data(value[ITEM_EMISSION_TIME].asDouble()));
 	m_ps->SetValue(PS_COUNT, UICallback::Data(value[ITEM_COUNT].asInt()));
-	m_ps->SetValue(PS_FADEOUT_TIME, UICallback::Data(value[ITEM_FADEOUT_TIME].asInt()));
-	m_ps->SetValue(PS_LIFE, UICallback::Data(value[ITEM_LIFE][ITEM_ATTR_CENTER].asInt(), value[ITEM_LIFE][ITEM_ATTR_OFFSET].asInt()));
+	m_ps->SetValue(PS_FADEOUT_TIME, UICallback::Data(value[ITEM_FADEOUT_TIME].asDouble()));
+	m_ps->SetValue(PS_LIFE, UICallback::Data(value[ITEM_LIFE][ITEM_ATTR_CENTER].asDouble(), value[ITEM_LIFE][ITEM_ATTR_OFFSET].asDouble()));
 	m_ps->SetValue(PS_POSITION, UICallback::Data(value[ITEM_POSITION][ITEM_ATTR_X].asInt(), value[ITEM_POSITION][ITEM_ATTR_Y].asInt()));
 	m_ps->SetValue(PS_DIRECTION, UICallback::Data(value[ITEM_DIRECTION][ITEM_ATTR_CENTER].asInt(), value[ITEM_DIRECTION][ITEM_ATTR_OFFSET].asInt()));
-	m_ps->SetValue(PS_SCALE, UICallback::Data(value[ITEM_SCALE][ITEM_ATTR_START].asInt(), value[ITEM_SCALE][ITEM_ATTR_END].asInt()));
+	m_ps->SetValue(PS_SCALE, UICallback::Data(value[ITEM_SCALE][ITEM_ATTR_START].asDouble(), value[ITEM_SCALE][ITEM_ATTR_END].asDouble()));
 	m_ps->SetValue(PS_SPEED, UICallback::Data(value[ITEM_SPEED][ITEM_ATTR_CENTER].asInt(), value[ITEM_SPEED][ITEM_ATTR_OFFSET].asInt()));
 	m_ps->SetValue(PS_GRAVITY, UICallback::Data(value[ITEM_GRAVITY][ITEM_ATTR_CENTER].asInt(), value[ITEM_GRAVITY][ITEM_ATTR_OFFSET].asInt()));
 	m_ps->SetValue(PS_RADIAL_ACC, UICallback::Data(value[ITEM_RADIAL_ACC][ITEM_ATTR_CENTER].asInt(), value[ITEM_RADIAL_ACC][ITEM_ATTR_OFFSET].asInt()));
 	m_ps->SetValue(PS_TANGENTIAL_ACC, UICallback::Data(value[ITEM_TANGENTIAL_ACC][ITEM_ATTR_CENTER].asInt(), value[ITEM_TANGENTIAL_ACC][ITEM_ATTR_OFFSET].asInt()));
 	m_ps->SetValue(PS_COS_AMPLITUDE, UICallback::Data(value[ITEM_COS][ITEM_COS_AMPLITUDE][ITEM_ATTR_CENTER].asInt(), value[ITEM_COS][ITEM_COS_AMPLITUDE][ITEM_ATTR_OFFSET].asInt()));
-	m_ps->SetValue(PS_COS_FREQUENCY, UICallback::Data(value[ITEM_COS][ITEM_COS_FREQUENCY][ITEM_ATTR_CENTER].asInt(), value[ITEM_COS][ITEM_COS_FREQUENCY][ITEM_ATTR_OFFSET].asInt()));
+	m_ps->SetValue(PS_COS_FREQUENCY, UICallback::Data(value[ITEM_COS][ITEM_COS_FREQUENCY][ITEM_ATTR_CENTER].asDouble(), value[ITEM_COS][ITEM_COS_FREQUENCY][ITEM_ATTR_OFFSET].asDouble()));
 
 	wxString dir = d2d::FilenameTools::getFileDir(m_filepath) + "\\";
 	wxString path = d2d::FilenameTools::getAbsolutePath(dir, value["symbol_path"].asString());
